fixing InstallController

+++ protected/controllers/InstallController.php
@@ -12,16 +12,18 @@ class InstallController extends Controller {
             $model->attributes = $_POST['AdminSetup'];
             Setting::set('db', $model->attributes);
             if ($this->validateDB()) {
-                $sql_set_db = "USE `".Setting::get('db.dbname')."`;";
-                $sql_content = file_get_contents(Setting::getRootPath() . '/installer/database/plansys.sql');
-                $sql_content = $sql_set_db.' '.$sql_content;
-                
-                $conn = Setting::getDB();
-                $db = new PDO($conn['connectionString'], $conn['username'], $conn['password']);
-                $command = $db->prepare($sql_content);
-                $command->execute();
-                
-                unlink(Setting::getRootPath() . '/installer/setup_db.lock');
+                if(file_exists(Setting::getRootPath() . '/installer/setup_db.lock')){
+                    $sql_set_db = "USE `".Setting::get('db.dbname')."`;";
+                    $sql_content = file_get_contents(Setting::getRootPath() . '/installer/database/plansys.sql');
+                    $sql_content = $sql_set_db.' '.$sql_content;
+
+                    $conn = Setting::getDB();
+                    $db = new PDO($conn['connectionString'], $conn['username'], $conn['password']);
+                    $command = $db->prepare($sql_content);
+                    $command->execute();
+                    
+                    unlink(Setting::getRootPath() . '/installer/setup_db.lock');
+                }
                 $this->redirect(array('site/login'));
             } else {
                 $model->errors = array(
